changed routes according to new folder structure in backend

+++ src/api/UserAdministrationAPI.jsx
@@ -3,7 +3,7 @@ import axiosInstanceAPI from './axiosInstanceAPI';
 // Function to get users from the backend
 export const getDataRoles = async () => {
   try {
-    const response = await axiosInstanceAPI.get('/getUserData.php');
+    const response = await axiosInstanceAPI.get('/userAdministration/getUserData.php');
     return response.data; // Return the fetched data
   } catch (error) {
     return null; // Return null or handle the error appropriately
@@ -12,7 +12,7 @@ export const getDataRoles = async () => {
 
 export const deleteUserFromDB = async (id, toast, data, setData) => {
   try {
-    await axiosInstanceAPI.delete(`/deleteUser.php?id=${id}`);
+    await axiosInstanceAPI.delete(`/userAdministration/deleteUser.php?id=${id}`);
     setData(data.filter(employee => employee.id !== id));
     toast({
       description: "Mitarbeiter wurde erfolgreich gelöscht.",
@@ -37,7 +37,7 @@ export const deleteUserFromDB = async (id, toast, data, setData) => {
 export const updateUser = async (id, formData, setLoading, setData, toast, getDataRoles) => {
   setLoading(true);
   try {
-    const response = await axiosInstanceAPI.post(`/updateUser.php`, formData, {
+    const response = await axiosInstanceAPI.post(`/userAdministration/updateUser.php`, formData, {
       headers: {
         'Content-Type': 'application/json',
       },
